mypage-password-view

+++ app/controllers/goods_controller.rb
@@ -11,8 +11,7 @@ class GoodsController < ApplicationController
   end
 
   def create
-    binding.pry
-    @good = Good.new(good_params)
+    @good = Good.create(good_params)
     if @good.save
       params[:images][:image].each do |image|
         Image.create(image: image, good_id: @good.id )
@@ -20,9 +19,8 @@ class GoodsController < ApplicationController
       params[:category_ids][:category_id].each do |category_id|
         CategoryGood.create(category_id: category_id, good_id: @good.id)
       end
-      redirect_to root_path
     end
-
+    redirect_to root_path
   end
 
   def destroy
@@ -36,26 +34,6 @@ class GoodsController < ApplicationController
     gon.good = @good
     gon.good_images = @good.images
 
-    require 'base64'
-    require 'fog-aws'
-
-    gon.good_images_binary_datas = []
-    if Rails.env.production?
-      client = Aws::S3::Client.new(
-        region: 'ap-northeast-1',
-        access_key_id: Rails.application.credentials.aws[:access_key_id],
-        secret_access_key: Rails.application.credentials.aws[:secret_access_key],
-      )
-      @good.images.each do |image|
-        binary_data = client.get_object(bucket: 'freemarket-sample-64i-image', key: image.image_url.file.path).body.read
-        gon.good_images_binary_datas << Base64.strict_encode64(binary_data)
-      end
-    else
-      @good.images.each do |image|
-        binary_data = File.read(image.image.file.file)
-        gon.good_images_binary_datas << Base64.strict_encode64(binary_data)
-      end
-    end
   end
 
 
@@ -67,17 +45,21 @@ class GoodsController < ApplicationController
 
 
   def update
-    
     @good = Good.find(params[:id])
     @good.update(good_params)
     if @good.update(good_params)
-      images = Image.where( good_id: @good.id )
-      images.each do |image|
-        image.destroy
+      if params[:images][:image]
+        params[:images][:image].each do |image|
+          Image.create(image: image, good_id: @good.id)
+        end
       end
-      params[:images][:image].each do |image|
-        Image.create(image: image, good_id: @good.id)
+      if params[:destroy][:ids]
+        params[:destroy][:ids].each do |id|
+         image = Image.find(id)
+         image.destroy
+        end
       end
+      
       categories = CategoryGood.where( good_id: @good.id)
       categories.each do |category|
         category.destroy
@@ -86,12 +68,12 @@ class GoodsController < ApplicationController
         CategoryGood.create(category_id: category_id, good_id: @good.id)
       end
     end
-      
+    redirect_to good_path(@good.id)
   end
 
 
   private
-  
+
   def good_params
     params.require(:good).permit(:name, :explain, :size, :price, :method, :ship, :burden, 
       :status, :brand_id, :area_id, :user_id, images_attribute: [:image], category_ids: [])
@@ -100,8 +82,7 @@ class GoodsController < ApplicationController
   def move_to_sign_up
     redirect_to new_user_session_path unless user_signed_in?
   end
-
-
+  
 end
